requirements.txt
django==1.10
django-compressor
django-extensions
django-libsass
django-model-utils
django-redis
psycopg2
redis
django-kdl-ldap
django-haystack==2.7.dev0
Mezzanine
Pillow
argparse
bleach
django-registration
filebrowser-safe
grappelli-safe
future
html5lib
lxml
ordereddict
pyproj
pysolr
pytz
requests
six
weasyprint
kdu-jp2